requirements.txt
Django==4.1.3  
djangorestframework==3.13.1  
drf-yasg==1.20.0 
Pillow==10.1
